refactor: add templates for object and array
