+++ packages/dm-core-plugins/src/form/fields/ObjectField.tsx
@@ -1,275 +1,24 @@
 import {
   EBlueprint,
-  EntityPickerDialog,
-  EntityView,
-  ErrorResponse,
   getKey,
   Loading,
-  resolveRelativeAddress,
-  splitAddress,
-  TLinkReference,
   useBlueprint,
-  useDMSS,
-  useDocument,
 } from '@development-framework/dm-core'
-import { Typography } from '@equinor/eds-core-react'
-import { add, chevron_down, chevron_up, edit } from '@equinor/eds-icons'
-import { AxiosError } from 'axios'
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import { useFormContext } from 'react-hook-form'
-import TooltipButton from '../../common/TooltipButton'
-import { OpenObjectButton } from '../components/OpenObjectButton'
-import { useRegistryContext } from '../context/RegistryContext'
 import { getWidget } from '../context/WidgetContext'
-import { Fieldset, Legend } from '../styles'
-import { TContentProps, TObjectFieldProps, TUiRecipeForm } from '../types'
-import RemoveObject from '../components/RemoveObjectButton'
-import AddObject from '../components/AddObjectButton'
+import { TField, TUiRecipeForm } from '../types'
 import { defaultConfig } from '../components/Form'
 import { getDisplayLabel } from '../utils/getDisplayLabel'
+import { ObjectStorageUncontainedTemplate } from '../templates/ObjectStorageUncontainedTemplate'
+import { ObjectModelContainedTemplate } from '../templates/ObjectModelContainedTemplate'
+import { ObjectModelUncontainedTemplate } from '../templates/ObjectModelUncontainedTemplate'
 
-const SelectReference = (props: {
-  attributeType: string
-  namePath: string
-}) => {
-  const [showModal, setShowModal] = useState<boolean>(false)
-  const { setValue, watch } = useFormContext()
-  const dmssAPI = useDMSS()
-  const { idReference } = useRegistryContext()
-  const value = watch(props.namePath)
-
-  const onChange = (address: string) => {
-    const reference: TLinkReference = {
-      type: EBlueprint.REFERENCE,
-      referenceType: 'link',
-      address: address,
-    }
-    const options = {
-      shouldValidate: true,
-      shouldDirty: true,
-      shouldTouch: true,
-    }
-
-    const request = value
-      ? dmssAPI.documentUpdate({
-          idAddress: `${idReference}.${props.namePath}`,
-          data: JSON.stringify(reference),
-        })
-      : dmssAPI.documentAdd({
-          address: `${idReference}.${props.namePath}`,
-          document: JSON.stringify(reference),
-        })
-    request
-      .then(() => {
-        setValue(props.namePath, reference, options)
-      })
-      .catch((error: AxiosError<ErrorResponse>) => {
-        console.error(error)
-      })
-  }
-
-  return (
-    <>
-      <TooltipButton
-        title={`${value ? 'Edit' : 'Add'} and save`}
-        button-variant="ghost_icon"
-        button-onClick={() => setShowModal(true)}
-        icon={value ? edit : add}
-      />
-      <EntityPickerDialog
-        data-testid={`select-${props.namePath}`}
-        onChange={onChange}
-        typeFilter={props.attributeType}
-        showModal={showModal}
-        setShowModal={setShowModal}
-      />
-    </>
-  )
-}
-
-export const StorageUncontainedAttribute = (props: TContentProps) => {
-  const { namePath, uiRecipe, attribute, uiAttribute } = props
-  const { watch, setValue } = useFormContext()
-  const { idReference, onOpen } = useRegistryContext()
-  const { dataSource, documentPath } = splitAddress(idReference)
-  const { config } = useRegistryContext()
-  const value = watch(namePath)
-  const address = resolveRelativeAddress(
-    value.address,
-    documentPath,
-    dataSource
-  )
-  const { document, isLoading } = useDocument<any>(address, 1)
-
-  useEffect(() => {
-    if (!isLoading && document) setValue(namePath, document)
-  }, [document])
-
-  return (
-    <Fieldset>
-      <Legend>
-        <Typography bold={true}>{getDisplayLabel(attribute)}</Typography>
-        {!config.readOnly && (
-          <SelectReference
-            attributeType={attribute.attributeType}
-            namePath={namePath}
-          />
-        )}
-        {attribute.optional && address && !config.readOnly && (
-          <RemoveObject address={address} namePath={namePath} />
-        )}
-        {address && onOpen && !uiAttribute?.showInline && (
-          <OpenObjectButton
-            viewId={namePath}
-            viewConfig={{
-              type: 'ReferenceViewConfig',
-              scope: '',
-              recipe: uiRecipe?.name,
-            }}
-            idReference={address}
-          />
-        )}
-      </Legend>
-      {address && !(onOpen && !uiAttribute?.showInline) && (
-        <EntityView
-          idReference={address}
-          type={attribute.attributeType}
-          recipeName={uiRecipe?.name}
-          onOpen={onOpen}
-        />
-      )}
-    </Fieldset>
-  )
-}
-
-export const ContainedAttribute = (
-  props: TContentProps
-): React.ReactElement => {
-  const { namePath, uiAttribute, uiRecipe, attribute } = props
-  const { watch } = useFormContext()
-  const { idReference, onOpen, config } = useRegistryContext()
-
-  const [isExpanded, setIsExpanded] = useState(
-    uiAttribute?.showExpanded !== undefined
-      ? uiAttribute?.showExpanded
-      : config.showExpanded
-  )
-  const value = watch(namePath)
-  const isDefined = value && Object.keys(value).length > 0
-  return (
-    <Fieldset>
-      <Legend>
-        <Typography bold={true}>{getDisplayLabel(attribute)}</Typography>
-        {attribute.optional &&
-          !config.readOnly &&
-          (isDefined ? (
-            <RemoveObject namePath={namePath} />
-          ) : (
-            <AddObject
-              namePath={namePath}
-              type={attribute.attributeType}
-              defaultValue={attribute.default}
-            />
-          ))}
-        {isDefined && !(onOpen && !uiAttribute?.showInline) && (
-          <TooltipButton
-            title="Expand"
-            button-variant="ghost_icon"
-            button-onClick={() => setIsExpanded(!isExpanded)}
-            icon={isExpanded ? chevron_up : chevron_down}
-          />
-        )}
-        {isDefined && onOpen && !uiAttribute?.showInline && (
-          <OpenObjectButton
-            viewId={namePath}
-            idReference={idReference}
-            viewConfig={{
-              type: 'ReferenceViewConfig',
-              scope: namePath,
-              recipe: uiRecipe?.name,
-            }}
-          />
-        )}
-      </Legend>
-      {isDefined && !(onOpen && !uiAttribute?.showInline) && isExpanded && (
-        <EntityView
-          recipeName={uiRecipe.name}
-          idReference={`${idReference}.${namePath}`}
-          type={attribute.attributeType}
-          onOpen={onOpen}
-        />
-      )}
-    </Fieldset>
-  )
-}
-
-export const UncontainedAttribute = (
-  props: TContentProps
-): React.ReactElement => {
-  const { namePath, uiAttribute, uiRecipe, attribute } = props
-  const { watch } = useFormContext()
-  const { idReference, onOpen, config } = useRegistryContext()
-  const [isExpanded, setIsExpanded] = useState(
-    uiAttribute?.showExpanded !== undefined
-      ? uiAttribute?.showExpanded
-      : config.showExpanded
-  )
-  const value = watch(namePath)
-  const { dataSource, documentPath } = splitAddress(idReference)
-  const address =
-    value && value.address && value.referenceType === 'link'
-      ? resolveRelativeAddress(value.address, documentPath, dataSource)
-      : undefined
-  return (
-    <Fieldset>
-      <Legend>
-        <Typography bold={true}>{getDisplayLabel(attribute)}</Typography>
-        {!config.readOnly && (
-          <SelectReference
-            attributeType={attribute.attributeType}
-            namePath={namePath}
-          />
-        )}
-        {attribute.optional && address && !config.readOnly && (
-          <RemoveObject namePath={namePath} />
-        )}
-        {address && !(onOpen && !uiAttribute?.showInline) && (
-          <TooltipButton
-            title="Expand"
-            button-variant="ghost_icon"
-            button-onClick={() => setIsExpanded(!isExpanded)}
-            icon={isExpanded ? chevron_up : chevron_down}
-          />
-        )}
-        {address && onOpen && !uiAttribute?.showInline && (
-          <OpenObjectButton
-            viewId={namePath}
-            viewConfig={{
-              type: 'ReferenceViewConfig',
-              scope: '',
-              recipe: uiRecipe?.name,
-            }}
-            idReference={address}
-          />
-        )}
-      </Legend>
-      {address && !(onOpen && !uiAttribute?.showInline) && isExpanded && (
-        <EntityView
-          idReference={address}
-          type={attribute.attributeType}
-          recipeName={uiRecipe?.name}
-          onOpen={onOpen}
-        />
-      )}
-    </Fieldset>
-  )
-}
-
-export const ObjectField = (props: TObjectFieldProps): React.ReactElement => {
+export const ObjectField = (props: TField): React.ReactElement => {
   const { namePath, uiAttribute, attribute } = props
   const { getValues } = useFormContext()
   const values = getValues(namePath)
-  const valuesIsStorageReference =
+  const isStorageUncontained =
     values !== undefined &&
     'referenceType' in values &&
     values['referenceType'] === 'storage'
@@ -281,15 +30,15 @@ export const ObjectField = (props: TObjectFieldProps): React.ReactElement => {
         {...props}
         label={getDisplayLabel(attribute)}
         onChange={() => null}
-        id={valuesIsStorageReference ? values['address'] : namePath}
+        id={isStorageUncontained ? values['address'] : namePath}
         // if the attribute type is an object, we need to find the correct type from the values.
       />
     )
   }
 
   return (
-    <ObjectTypeSelector
-      uiAttribute={uiAttribute}
+    <ObjectTemplateSelector
+      namePath={namePath}
       attribute={{
         ...attribute,
         attributeType:
@@ -297,14 +46,12 @@ export const ObjectField = (props: TObjectFieldProps): React.ReactElement => {
             ? values.type
             : attribute.attributeType,
       }}
-      namePath={namePath}
+      uiAttribute={uiAttribute}
     />
   )
 }
 
-export const ObjectTypeSelector = (
-  props: TObjectFieldProps
-): React.ReactElement => {
+export const ObjectTemplateSelector = (props: TField): React.ReactElement => {
   const { namePath, uiAttribute, attribute } = props
   const { blueprint, uiRecipes, isLoading, error } = useBlueprint(
     attribute.attributeType
@@ -317,7 +64,7 @@ export const ObjectTypeSelector = (
       `Failed to fetch blueprint for '${attribute.attributeType}'`
     )
   if (blueprint === undefined) return <div>Could not find the blueprint</div>
-  const attributeIsStorageReference =
+  const isStorageUncontained =
     value !== undefined &&
     'referenceType' in value &&
     value['referenceType'] === 'storage'
@@ -343,14 +90,16 @@ export const ObjectTypeSelector = (
       `No UiRecipe named "${uiRecipeName}" could be found for type "${attribute.attributeType}"`
     )
 
-  const Content = attributeIsStorageReference
-    ? StorageUncontainedAttribute
-    : attribute.contained ?? true
-    ? ContainedAttribute
-    : UncontainedAttribute
+  const isModelContained = attribute.contained ?? true
+
+  const Template = isStorageUncontained
+    ? ObjectStorageUncontainedTemplate
+    : isModelContained
+    ? ObjectModelContainedTemplate
+    : ObjectModelUncontainedTemplate
 
   return (
-    <Content
+    <Template
       attribute={attribute}
       namePath={namePath}
       blueprint={blueprint}
