+++ packages/dm-core-plugins/src/form/utils/isPrimitiveType.tsx
@@ -1,0 +1,3 @@
+export const isPrimitiveType = (value: string): boolean => {
+  return ['string', 'number', 'integer', 'boolean'].includes(value)
+}
